+++ ChangeLog.md
@@ -5,8 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
-### Changed
+### Added
 - Support for morph targets / blend shapes (#8)
+- Support for animated morph targets / blend shapes
+### Changed
 - glTF importer inspector
   - Removed "Node Name Method" option from glTF importer inspector. It still an option at run-time, but is always `OriginalUnique` at design-time imports.
   - `Animation` setting is disabled if built-in package animation is disabled